Always use vendored version of fmt
tweedledum is incompatible with the latest version of fmt (11.2.0).
However, when a system-provided version of the library is discovered
through pkg-config, it is preferred over the vendored version, even when
the system package is too new.

This commit changes it to always use the vendored version, which is
known to be compatible. If it is still desired to use a system package,
an option can be added like the one removed in
0a01475057aeab5e7ca9ceed838719e51769e326.

It may be worth making the same change for Eigen3, nlohmann_json and
phmap. These also currently prefer a system package over the vendored
version, which may cause similar problems in future.